requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.3
pyaml==23.7.0
pydantic==2.1.1
pydantic_core==2.4.0
PyYAML==6.0.1
sniffio==1.3.0
typing_extensions==4.7.1
uvicorn~=0.23.0
websocket-client~=1.6.2
docker~=7.0.0
fastapi~=0.103.0
setuptools~=68.0.0
requests~=2.31.0
yaml~=0.2.5
kubernetes~=29.0.0
pandas~=1.5.3
numpy~=1.23.5
plotly~=5.19.0
streamlit~=1.31.1
python-dotenv~=0.21.0